Align api of Duration with DataSize
Record DataSize unit so unit does not change with printing

+++ discovery/src/main/java/com/proofpoint/discovery/client/ServiceDescriptorsUpdater.java
@@ -138,7 +138,7 @@ public final class ServiceDescriptorsUpdater
             {
                 refresh();
             }
-        }, (long) delay.toMillis(), TimeUnit.MILLISECONDS);
+        }, delay.toMillis(), TimeUnit.MILLISECONDS);
     }
 
     private static <V> ListenableFuture<V> chainedCallback(
